+++ clang/lib/CodeGen/ItaniumCXXABI.cpp
@@ -3196,6 +3196,9 @@ void ItaniumRTTIBuilder::BuildVTablePointer(const Type *Ty) {
   case Type::Pipe:
     llvm_unreachable("Pipe types shouldn't get here");
 
+  case Type::ExtInt:
+    llvm_unreachable("Ext Int Types shouldn't get here");
+
   case Type::Builtin:
   // GCC treats vector and complex types as fundamental types.
   case Type::Vector:
@@ -3451,6 +3454,9 @@ llvm::Constant *ItaniumRTTIBuilder::BuildTypeInfo(
 
   case Type::Pipe:
     llvm_unreachable("Pipe type shouldn't get here");
+
+  case Type::ExtInt:
+    llvm_unreachable("Ext Int type shouldn't get here");
 
   case Type::ConstantArray:
   case Type::IncompleteArray: